Add the fancy multi level FAB

+++ app/src/main/java/cd/acgt/acgtexp/activites/BaseAddActivity.java
@@ -26,7 +26,7 @@ public class BaseAddActivity extends AppCompatActivity {
         int type = intent.getIntExtra(Constant.KEY_TYPE, Constant.PROPRIETE_TYPE);
         String codeProjet = intent.getStringExtra(Constant.KEY_CODE_PROJECT);
         long riverainID = intent.getLongExtra(Constant.KEY_CODE_RIVERAIN, 0L);
-        long proprieteID = intent.getLongExtra(Constant.KEY_CODE_PROPRIETE, 0L);
+        int proprieteID = intent.getIntExtra(Constant.KEY_CODE_PROPRIETE, 0);
         Toast.makeText(this, "" + codeProjet, Toast.LENGTH_LONG).show();
         switch(type) {
             case Constant.PROPRIETE_TYPE:
